requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.34.0
accelerate>=0.23.0
flask>=2.3.0
flask-cors>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
opencv-python>=4.8.0
safetensors>=0.4.0
compel>=2.0.0
